# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API ایده‌ها",
        "description": "API برای مدیریت ایده‌ها",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "سرور اصلی"
        }
    ],
    "paths": {
        "/api/v1/admin/categories": {
            "get": {
                "tags": [
                    "Admin - categories"
                ],
                "summary": "لیست دسته‌بندی‌ها",
                "description": "لیست دسته‌بندی‌ها با قابلیت فیلتر بر اساس شناسه مجموعه و مرتب‌سازی بر اساس زمان ایجاد",
                "operationId": "categoriesList",
                "parameters": [
                    {
                        "name": "department_id",
                        "in": "query",
                        "description": "فیلتر بر اساس آیدی مجموعه",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "فیلد مرتب‌سازی",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "created_at"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "عملیات موفق",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "categories": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/CategoryIndexResource"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "لیست دسته‌بندی‌ها با موفقیت دریافت شد"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "خطای سرور",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "خطای داخلی سرور"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Admin - categories"
                ],
                "summary": "ایجاد دسته‌بندی جدید",
                "description": "ایجاد یک دسته‌بندی جدید در سیستم",
                "operationId": "storeCategory",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "title",
                                    "status",
                                    "created_by",
                                    "updated_by",
                                    "department_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "دسته‌بندی جدید"
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "دسته‌بندی جدید"
                                    },
                                    "status": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "created_by": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "updated_by": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "department_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "توضیحات دسته‌بندی",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "دسته‌بندی با موفقیت ایجاد شد",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "category": {
                                                    "$ref": "#/components/schemas/CategoryStoreResource"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "دسته‌بندی با موفقیت ایجاد شد"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "خطای اعتبارسنجی",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "خطا در اعتبارسنجی داده‌ها"
                                        },
                                        "errors": {
                                            "properties": {
                                                "name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "نام دسته‌بندی الزامی است"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "خطای سرور",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "خطای داخلی سرور"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CategoryIndexResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "فناوری اطلاعات"
                    },
                    "department_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "status": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date",
                        "example": "1402/01/01"
                    }
                },
                "type": "object"
            },
...
            "CategoryStoreResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 5
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter your JWT Bearer token in the format **Bearer {your token}**"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}